Fix comment on --doing
I remember putting in this condition for ~!Local,createNew!~, and I
remember why. But I must have removed it at some point when it was no
longer needed, because it doesn't seem to come from anywhere now, and
was causing this bug after the recent code cleanup.

I can't see any reason to keep it, so I've removed it.